requirements.txt
discord.py
requests
dotenv